requirements.txt
beautifulsoup4==4.12.2
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.1
Flask-Mail==0.9.1
importlib-metadata==6.6.0
inexactsearch==1.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
nltk==3.8.1
numpy==1.22.0
pyspellchecker==0.7.2
regex==2023.3.23
silpa-common==0.3
simplemma==0.9.1
soundex==1.1.3
soupsieve==2.4.1
stop-words==2018.7.23
tqdm==4.65.0
Werkzeug==2.3.1
zipp==3.15.0
